requirements.txt
Django==3.0.6
djangorestframework==3.11.0
firebase-admin==4.3.0
python-decouple==3.3